# docs/swagger.yaml
basePath: /api
definitions:
  cmd_api_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.PasswordResetConfirm:
    properties:
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  models.PasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        minLength: 1
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  models.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
  github_com_AbhishekCS3459_find-me-backend_cmd_api_queue.Job:
    properties:
      created_at:
        type: string
      id:
        type: string
      max_retries:
        type: integer
      payload:
        items:
          type: integer
        type: array
      retries:
        type: integer
      type:
        type: string
    type: object
  github_com_AbhishekCS3459_find-me-backend_cmd_api_queue.QueueInfo:
    properties:
      jobs:
        description: First 10 jobs (if peek=true)
        items:
          $ref: '#/definitions/github_com_AbhishekCS3459_find-me-backend_cmd_api_queue.Job'
        type: array
      peeked:
        type: boolean
      stats:
        $ref: '#/definitions/github_com_AbhishekCS3459_find-me-backend_cmd_api_queue.QueueStats'
    type: object
  github_com_AbhishekCS3459_find-me-backend_cmd_api_queue.QueueStats:
    properties:
      key:
        description: Redis key name
        type: string
      length:
        description: Number of jobs waiting
        type: integer
      queue_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Abhishek Kumar Vema
    url: https://github.com/AbhishekCS3459
  description: |-
    Find Me backend API with PostgreSQL, JWT authentication, and role-based access.
    **Features:**
    - Clean Architecture (Handler-Service-Repository)
    - JWT Authentication
    - Database Migrations
    - OpenAPI Documentation
  title: Find Me API
  version: 1.0.0
paths:
  /api/admin/queues:
    get:
      consumes:
      - application/json
      description: Get statistics for all queues (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all queues
      tags:
      - queue
  /api/admin/queues/{queueName}:
    get:
      consumes:
      - application/json
      description: Get statistics for a specific queue (admin only)
      parameters:
      - description: Queue name
        in: path
        name: queueName
        required: true
        type: string
      - description: 'Peek at jobs (default: false)'
        in: query
        name: peek
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_AbhishekCS3459_find-me-backend_cmd_api_queue.QueueInfo'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get queue statistics
      tags:
      - queue
  /api/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change password for the authenticated user
      parameters:
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Authentication
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /api/auth/request-password-reset:
    post:
      consumes:
      - application/json
      description: Request a password reset token via email
      parameters:
      - description: Password reset request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      summary: Request password reset
      tags:
      - Authentication
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using a reset token
      parameters:
      - description: Password reset confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PasswordResetConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      summary: Reset password
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"